requirements.txt
PyGithub==2.1.1
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.6.1
python-gitlab==4.4.0  # For future GitLab support
azure-devops==7.1.0b4  # For future Azure DevOps support
GitPython==3.1.41  # For local repository operations